requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.4.0
pymongo==4.7.2
python-dotenv==1.0.1
pymupdf==1.24.3
langchain==0.2.5
langchain-google-genai==1.0.6
langchain-pinecone==0.1.1
pinecone-client==4.1.0
google-generativeai==0.7.2
pydantic[email]==2.7.4